requirements.txt
aiofiles==0.8.0
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
flake8==4.0.1
flake8-commas==2.1.0
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
frozenlist==1.3.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
mccabe==0.6.1
multidict==6.0.2
mypy==0.931
mypy-extensions==0.4.3
pep8-naming==0.12.1
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
rfc3986==1.5.0
sniffio==1.2.0
tomli==2.0.1
typing_extensions==4.1.1
yarl==1.7.2
